requirements.txt
fastapi
uvicorn
requests
langchain_google_genai
langchain_community
rapidfuzz
pypdf
faiss-cpu
langchain_experimental
httpx[http2]
pymupdf
numpy